Ændret i hvordan uge bliver hånteret mellem Reservations komponenter
